[Routing page]Step 1: Added all routing functions in a separate file and imported them. Status: Everything works!
